pdf-vectorizer/pyproject.toml
[project]
name = "pdf-vectorizer"
version = "0.1.0"
description = "PDF text extraction, chunking, and vectorization pipeline"
requires-python = ">=3.11"
dependencies = [
    "pymupdf>=1.25.1",
    "pdfplumber>=0.11.0",
    "langchain>=0.3.14",
    "langchain-text-splitters>=0.3.2",
    "sentence-transformers>=3.3.1",
    "pandas>=2.3.3",
    "numpy>=2.2.1",
    "pyyaml>=6.0.3",
    "python-dotenv>=1.2.1",
    "rich>=13.0.0",
]